feat: add debugger to linear search
