+++ docs/1-search-algorithms/1-linear-search.mdx
@@ -2,7 +2,7 @@
 sidebar_position: 1
 ---
 
-import { LinearArraySearch } from '../../src/components/Algorithms/linear-array-search'
+import { LinearArraySearch } from '../../src/algorithms/code/linear-array-search/linear-array-search'
 
 # Linear search
