+++ docs/1-search-algorithms/2-binary-search.mdx
@@ -2,7 +2,7 @@
 sidebar_position: 2
 ---
 
-import { BinaryArraySearch } from '../../src/components/Algorithms/binary-array-search'
+import { BinaryArraySearch } from '../../src/algorithms/code/binary-array-search/binary-array-search'
 
 # Binary search
