+++ src/algorithms/code/styles.module.css
@@ -1,0 +1,4 @@
+.boxContainer {
+  display: flex;
+  flex-wrap: wrap;
+}
