+++ src/algorithms/ui/debugger/styles.module.css
@@ -1,0 +1,5 @@
+.container {
+  p {
+    margin-bottom: 0;
+  }
+}
