--- src/components/Algorithms/styles.module.css
@@ -1,7 +1,0 @@
-.active {
-  background-color: var(--ifm-color-primary);
-}
-
-.found {
-  background-color: var(--ifm-color-success);
-}
